Improvements for PR petele/harmony #4
Add a time.sleep(0.5) to send_command in client.py to give the command a
chance to complete. This is an ugly hack, we need to actually follow the
protocol, but works fine for now.

Support using device labels or ids case-insensitively with --device.

Give an error if the device does not exist.

Update README.md.